+++ resources/views/partials/nav.blade.php
@@ -8,7 +8,9 @@
         </button>
         <div class="collapse navbar-collapse" id="navbarSupportedContent">
             <ul class="navbar-nav mr-auto">
-
+                <li class="nav-item">
+                    <a class="nav-link" href="{{ route('patients.index') }}">Patients</a>
+                </li>
             </ul>
             <ul class="navbar-nav ml-auto">
                 @auth