+++ resources/views/patients/create.blade.php
@@ -1,0 +1,51 @@
+@extends('layouts.backend')
+
+@section('title', 'Create New Patient')
+
+@section('content')
+
+<div class="container">
+    <div class="row">
+        <div class="col-md-12">
+            <div class="card">
+                <div class="card-header">Create New Patient</div>
+                <div class="card-body">
+                    <form method="POST" action="{{ route('patients.store') }}">
+                        @csrf
+                        <div class="form-row">
+                            <div class="form-group col-md-4">
+                                <label for="name">Name</label>
+                                <input type="text" class="form-control" name="name" id="name" required autofocus>
+                            </div>
+                            <div class="form-group col-md-4">
+                                <label for="email">Email</label>
+                                <input type="email" class="form-control" name="email" id="email">
+                            </div>
+                            <div class="form-group col-md-4">
+                                <label for="phone">Phone</label>
+                                <input type="text" class="form-control" name="phone" id="phone">
+                            </div>
+                        </div>
+                        <div class="form-row">
+                            <div class="form-group col-md-4">
+                                <label for="gender">Gender</label>
+                                <select id="gender" name="gender" class="form-control" required>
+                                    <option value="" selected></option>
+                                    <option value="male">Male</option>
+                                    <option value="female">Female</option>
+                                </select>
+                            </div>
+                            <div class="form-group col-md-4">
+                                <label for="dob">Birth Date</label>
+                                <input type="date" class="form-control" name="dob" id="dob">
+                            </div>
+                        </div>
+                        <button type="submit" class="btn btn-primary">Save</button>
+                    </form>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+
+@endsection