+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 0.0.3 2019-11-11
+
+### Added
+- Add patients module
+
+### Changed
+- Change login attribute from email to username
+- Add soft deleting to user model
+
+## 0.0.2 2019-11-10
+
+### Added
+- Add base layout and child layouts
+
 ## 0.0.1 - 2019-11-08
 
 ### Added